# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.ProductCategory:
    enum:
    - Grocery
    - FruitAndVeg
    - Meat
    - Poultry
    - Seafood
    - Bakery
    - Liquor
    - GeneralStore
    type: string
    x-enum-varnames:
    - ProductCategoryGrocery
    - ProductCategoryFruitAndVeg
    - ProductCategoryMeat
    - ProductCategoryPoultry
    - ProductCategorySeafood
    - ProductCategoryBakery
    - ProductCategoryLiquor
    - ProductCategoryGeneralStore
...
  request.CreateAdminRequest:
    properties:
      establishment_id:
        type: integer
      is_active:
        type: boolean
      user_id:
        type: integer
    required:
    - establishment_id
    - user_id
    type: object
...
  request.CreateEstablishmentRequest:
    properties:
      address:
        type: string
      admin_id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      ruc:
        type: string
    required:
    - address
    - admin_id
    - name
    - phone
    - ruc
    type: object
...
  request.UpdateAdminRequest:
    properties:
      is_active:
        type: boolean
    type: object
...
  response.AdminResponse:
    properties:
      created_at:
        type: string
      establishment:
        $ref: '#/definitions/response.EstablishmentResponse'
      establishment_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.EstablishmentResponse:
    properties:
      address:
        type: string
      admin:
        $ref: '#/definitions/response.AdminResponse'
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      products:
        items:
          $ref: '#/definitions/response.ProductResponse'
        type: array
      ruc:
        type: string
      updated_at:
        type: string
    type: object
...
  response.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/entities.ProductCategory'
      created_at:
        type: string
      description:
        type: string
      establishment:
        $ref: '#/definitions/response.EstablishmentResponse'
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing finances in small businesses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Final Assignment Finance API Rest
  version: "1.0"
paths:
  /api/v1/admins:
    get:
      consumes:
      - application/json
      description: Gets a list of all admins.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.AdminResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all admins
      tags:
      - Admins
    post:
      consumes:
      - application/json
      description: Creates a new admin with the provided data
      parameters:
      - description: Admin data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/request.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.AdminResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new admin
      tags:
      - Admins
  /api/v1/admins/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an admin by their ID.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Admin deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid admin ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Admin not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete admin
      tags:
      - Admins
    get:
      consumes:
      - application/json
      description: Gets an admin by their ID.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AdminResponse'
        "400":
          description: Invalid admin ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Admin not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get admin by ID
      tags:
      - Admins
    put:
      consumes:
      - application/json
      description: Updates an admin\'s data.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated admin data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/request.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Admin updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid admin ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Admin not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update admin
      tags:
      - Admins
...
  /api/v1/establishments:
    get:
      description: Retrieve a list of all establishments.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.EstablishmentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all establishments
      tags:
      - Establishments
    post:
      consumes:
      - application/json
      description: Create a new establishment with admin details.
      parameters:
      - description: Establishment details
        in: body
        name: establishment
        required: true
        schema:
          $ref: '#/definitions/request.CreateEstablishmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.EstablishmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new establishment
      tags:
      - Establishments
...
swagger: "2.0"