# internal/model/entities/product_category.go
package entities

type ProductCategory string

const (
	ProductCategoryGrocery      ProductCategory = "Grocery"
	ProductCategoryFruitAndVeg  ProductCategory = "FruitAndVeg"
	ProductCategoryMeat         ProductCategory = "Meat"
	ProductCategoryPoultry      ProductCategory = "Poultry"
	ProductCategorySeafood      ProductCategory = "Seafood"
	ProductCategoryBakery       ProductCategory = "Bakery"
	ProductCategoryLiquor       ProductCategory = "Liquor"
	ProductCategoryGeneralStore ProductCategory = "GeneralStore"
)